mlops_hw1/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

PyYAML == 6.0
catboost == 1.1
marshmallow_dataclass == 8.5.9
matplotlib == 3.6.0
numpy == 1.23.2
pandas == 1.5.0
plotly == 5.11.0
scipy == 1.9.3
seaborn == 0.12.1
sklearn == 0.0
sweetviz == 2.1.4
